+++ R/get_predictions_clmm.R
@@ -16,7 +16,7 @@ get_predictions_clmm <- function(model, terms, value_adjustment, condition, ci.l
   fe.terms <- insight::find_predictors(model, flatten = TRUE)
 
   if (any(re.terms %in% names(values.at)) && !any(re.terms %in% fe.terms)) {
-    insight::format_warning(
+    insight::format_alert(
       "Predicted values can't be computed for levels of random effects from 'clmm' models.",
       sprintf(
         "Please remove following variables from 'terms': %s",